+++ NEWS.md
@@ -1,3 +1,9 @@
+# ggtangle 0.0.3
+
++ `drag_network()` from 'enrichplot' (2024-10-24, Thu)
++ `cnetplot()` method for 'list' (2024-10-24, Thu)
++ re-export `ggfun::td_filter()` and `ggplot2::geom_point()`
+
 # ggtangle 0.0.2
 
 + `geom_edge()` layer to draw network (2024-08-28, Wed)